Move component files to pages directry
